refactor: reorganize /sdn/util dir.
