+++ cmd/app/init.go
@@ -2,12 +2,10 @@ package cmd
 
 import (
 	"fmt"
-	"time"
 
 	"github.com/spf13/cobra"
 
 	"ws/dtn-satellite-sdn/sdn"
-	"ws/dtn-satellite-sdn/sdn/util"
 )
   
 
@@ -15,7 +13,6 @@ var (
 	url string
 	node int
 	interval int
-	is_count bool
 	is_test bool
 	version string
 
@@ -23,8 +20,6 @@ var (
 		Use:   "init",
 		Short: "Init a Satellite Network emulation environment.",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			startTimer := time.Now()
-
 			switch version {
 			case "v1":
 				if err := sdn.RunSatelliteSDN(url, node, interval); err != nil {
@@ -42,16 +37,6 @@ var (
 					}
 				}
 			}
-
-			// fmt.Printf("%v %v %v %v\n", tle, node, interval, is_test)
-
-			if is_count {
-				test_result, err := util.InitEnvTimeCounter(startTimer)
-				if err != nil {
-					return fmt.Errorf("count time failed: %v", err)
-				}
-				fmt.Printf("init Emulation Env Lasts For: %vs", test_result)
-			}
 			
 			return nil
 		},
@@ -62,7 +47,6 @@ func init() {
 	initCmd.Flags().StringVarP(&url, "url", "u", "", "v1: TLE file's path to read from / v2: The address of Position Calculation Module")
 	initCmd.Flags().IntVarP(&node, "node", "n", 3, "Expected node num")
 	initCmd.Flags().IntVarP(&interval, "interval", "i", -1, "Assign update interval for Satellite SDN Controller (-1 means 'no update')")
-	initCmd.Flags().BoolVar(&is_count, "count", false, "Open the time counter")
 	initCmd.Flags().BoolVar(&is_test, "test", false, "Open the test mode")
 	initCmd.Flags().StringVarP(&version, "version", "v", "", "SDN server's version")
